requirements.txt
Django==1.8.6
django-ckeditor==5.0.2
wheel==0.24.0
